src/main/java/com/group1/SDCapplication/login/jsonwebtoken/JwtTokenUtil.java
package com.group1.SDCapplication.login.jsonwebtoken;


import java.io.Serializable;
import java.util.*;
import java.util.function.Function;

import com.group1.SDCapplication.login.models.UserCredentials;
import com.group1.SDCapplication.models.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.stereotype.Component;
import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;


@Component
public class JwtTokenUtil implements Serializable {
    private static final long serialVersionUID = -2550185165626007488L;
    public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
    private String secret = "CSCI5308";
    public String getUsernameFromToken(String token) {
        return getClaimFromToken(token, Claims::getSubject);
    }
    public Date getExpirationDateFromToken(String token) {
        return getClaimFromToken(token, Claims::getExpiration);
    }
    public <T> T getClaimFromToken(String token, Function<Claims, T> claimsResolver) {
        final Claims claims = getAllClaimsFromToken(token);
        return claimsResolver.apply(claims);
    }
    public Claims getAllClaimsFromToken(String token) {
        return Jwts.parser().setSigningKey(secret).parseClaimsJws(token).getBody();
    }
    public Boolean isTokenExpired(String token) {
        final Date expiration = getExpirationDateFromToken(token);
        return expiration.before(new Date());
    }
    public String generateToken(User user) {

        return generateTokenWithRoles(user, new ArrayList<>());
    }
    public String generateTokenWithRoles(User user, List<String >roles ){
        Map<String, Object> claims = new HashMap<>();
        claims.put("UID",user.getId());
        claims.put("email",user.getEmail());
        claims.put("first name",user.getFirstname());
        claims.put("last name",user.getLastname());
        claims.put("roles", roles);
        return doGenerateToken(claims, "CSCI5308");
    }
    private String doGenerateToken(Map<String, Object> claims, String subject) {
        return Jwts.builder().setClaims(claims).setSubject(subject).setIssuedAt(new Date(System.currentTimeMillis()))
                .setExpiration(new Date(System.currentTimeMillis() + JWT_TOKEN_VALIDITY * 10))
                .signWith(SignatureAlgorithm.HS512, secret).compact();
    }
    public Boolean validateToken(String token, UserDetails userDetails) {
        final String username = getUsernameFromToken(token);
        return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
    }
}
